fix : duplicate test
